* Auto gen Operations (#1599)

* init

* ordered

* add feequoter to reduce overfitting

* fee quoter

* onramp/offramp

* standalone version, no gethwrappers

* gen to correct folder

* load from abi/bin

* cleanup

* minor fixes

* rm new ops

* rm issue with I-prefixed structs

* rm ping pong

* lint + go mod

* fix users

* reduce dups

* fix test

* make path more flexible

* readme

* implement all overloads

* force access modifier

* Init TAR in manual registration (#1605)

* Init TAR in manual registration

* Add support for lock release

* refactor init global config

* remove comment

* Fix some go lint errors

* Fix some

* Fix

* f

* Fix ports issue

* test 2 random ports

* feat: implement ConfigureTokenForTransfersSequence for the EVM tokens API (#1607)

* feat: implement ConfigureTokenForTransfersSequence for the EVM tokens API

* chore: fix typo in comment

* chore: update comment

* fix: rename sequences

* fix: add ChainAdapterVersion to ConfigureTokensForTransfers Apply test

* chore: update set-rate-limiter-config sequence name

* chore: rename

* fix: rate limits can be set by admin or owner

* chore: simplify ConfigureTokenForTransfer test

* chore: minor test improvements

* chore: minor test improvements

* feat: add support for token transfers to devenv (#1617)

* feat: implement ConfigureTokenForTransfersSequence for the EVM tokens API

* chore: fix typo in comment

* chore: update comment

* fix: rename sequences

* fix: add ChainAdapterVersion to ConfigureTokensForTransfers Apply test

* chore: update set-rate-limiter-config sequence name

* chore: rename

* fix: rate limits can be set by admin or owner

* chore: simplify ConfigureTokenForTransfer test

* chore: minor test improvements

* chore: minor test improvements

* feat: add support for token transfers to devenv

* chore: clean up

* Aw/rmn enablement (#1619)

* update the way we handle rmn

* hardcode rmnenabled in commit plugin

* tests

* add warn on hardcode rmnenabled

* changes, add warn on hardcode

* disable lane if rmn is enabled (#1627)

* avoid observing sequence numbers from chains that are rmn enabled

* lint

* fix test

* lint

* add unit tests, add defensive access, add comment

* lint

* Ccip-8552 import config from 1.5 and 1.6 contracts to be used for 1.7 deployment (#1626)

* updates

* updates

* more updates

* initial config import adapter

* fix build error

* review comments

* Generate all 1.6.1, 1.6.2, 1.6.4 operations (#1625)

* gen 1.6.1

* add 1.6.4

* generate shared base: TokenPool

* allow skipping deployment code

* fold 1.6.4 token pool into usdc_token_pool_cctp_v2

* gen cctp_message_transmitter_proxy

* add DO NOT EDIT warning

* fix RemotePoolModification

* fix PackageNameHyphen

* Implement full manual registration solana (#1621)

* Refactor Token Pool Ops

* Set authority tp

* Fix build

* Apply suggestions from code review


---------


* change uint32 to uint256 for maxFee param (#1637)

* Add support for view functions with params (#1632)

* add missing operation calls

* use ops over wrappers

* solana token pool expansion (#1623)

* Refactor Token Pool Ops

* Set authority tp

* Fix build

* tprl wip

* impls

* debugging

* move to devenv

* Apply suggestions from code review


* wip

* wip

* wip

* devenv e2e

* tests

* tests

* tests

* tests

* tests

* bug

* cr comments

---------


* Simplify cctpv1 pool (#1652)

* simplify CCTPv1 pool

* compilation fix

* test fix + remove USDCTokenPoolProxy not it use

* wrapper update

* add comment

* minor fixes

* add usdc_token_pool wrapper

* move _onlyXRamp to original place

* Create multisig for token solana (#1634)

* Refactor Token Pool Ops

* Set authority tp

* Fix build

* Apply suggestions from code review


* Create multisig for token solana

* Fix merge conflict

* Fix lint

* WIP

* Refactor method

* Refactor input

* Fix test

* Fix naming

* Fix test

* tests almost done

* WIP

* Disable tests

* Cmp pda

* Add execute proposals

* Fix reading PDAs

* Fix multisig

* Fix tests

* Fix go mod

* PR Feedback

---------


* usdc v1.6.5 ops (#1656)

* productionize token expansion API (#1647)

* add token metadata

* gomod

* wip

* compile

* compile

* tests

* compile

* compile

* Use token pool version instead of ChainAdapterVersion in ConfigureTokenForTransfers

* Bump chains/evm and chains/solana based on deployment pkg

* Bump ccv/chains/evm/deployment based on dependencies

---------

+++ chains/evm/contracts/test/mocks/MockE2EUSDCTokenMessenger.sol
@@ -84,7 +84,7 @@ contract MockE2EUSDCTokenMessenger is ITokenMessenger {
     bytes32 mintRecipient,
     address burnToken,
     bytes32 destinationCaller,
-    uint32 maxFee,
+    uint256 maxFee,
     uint32 minFinalityThreshold
   ) external {
     IBurnMintERC20(burnToken).transferFrom(msg.sender, address(this), amount);
